- Removed the downloader from the speechd module. - The downloaded has been moved to a seperate package, documentation has been added on how to use the downloader. - modules dot conf file can now set the models+voices search path.
